Made the Event Page
