+++ src/components/events/CategoryCard.js
@@ -4,13 +4,9 @@ import './event.css';
 export default class CategoryCard extends Component {
     render() {
         return (
-            <div className="categoryCardsGrid" onClick={this.props.onClick.bind(this, this.props.category.title, this.props.categoryIndex)}>
-                <div> 
-                    <img src={this.props.category.url} width="50px" height="50px" ></img>
-                </div>
-
+            <div className="categoryCardsGrid">
                 <div>
-                    <p style={{fontSize:"10px"}}>{this.props.category.title}</p>
+                    <p style={{fontSize:"15px"}}>{this.props.category}</p>
                 </div>
             </div>
         )
