+++ src/components/events/event.scss
@@ -3,9 +3,6 @@
     grid-template-rows:  0.1fr 1fr;
     text-align: center;
 }
-.backButton{
-    display: hidden;
-}
 
 //-----------------First Page----------------------
 .eventHeading{
