--- src/components/gallery/gallery.css
@@ -1,98 +1,0 @@
-#gallerySlider {
-  overflow: hidden;
-  position: relative;
-}
-
-#gallerySlider .galleryGrid {
-  display: -ms-grid;
-  display: grid;
-  -ms-grid-columns: (1fr)[10];
-      grid-template-columns: repeat(10, 1fr);
-  position: relative;
-  width: 1000%;
-  left: 0;
-  margin: 0;
-  background-color: #485461;
-  background-image: linear-gradient(315deg, #485461 0%, #28313b 74%);
-  -webkit-transition: 2s ease-in-out;
-  transition: 2s ease-in-out;
-  overflow: hidden;
-}
-
-.galleryGridComm {
-  text-align: center;
-  -webkit-transition: 0.5s ease-in-out;
-  transition: 0.5s ease-in-out;
-}
-
-.galleryGridComm .topText {
-  color: white;
-  position: absolute;
-  top: 20px;
-  z-index: 0;
-  font-size: 3rem;
-}
-
-.galleryGridComm .bottomText {
-  color: white;
-  position: absolute;
-  z-index: 0;
-  top: 90%;
-  right: 2vh;
-  font-size: 3rem;
-}
-
-.galleryGridComm img {
-  width: 100%;
-  height: 85vh;
-  -webkit-filter: grayscale(0%);
-          filter: grayscale(0%);
-  position: relative;
-  z-index: 1;
-  -webkit-transition: 1s cubic-bezier(0.215, 0.61, 0.355, 1);
-  transition: 1s cubic-bezier(0.215, 0.61, 0.355, 1);
-  border-radius: 10px;
-  -webkit-clip-path: polygon(0 0, 100% 0, 100% 100%, 0% 100%);
-          clip-path: polygon(0 0, 100% 0, 100% 100%, 0% 100%);
-}
-
-.galleryGridComm:hover img {
-  -webkit-filter: grayscale(60%);
-          filter: grayscale(60%);
-  -webkit-clip-path: polygon(40% 0, 100% 0, 58% 100%, 0% 100%);
-          clip-path: polygon(40% 0, 100% 0, 58% 100%, 0% 100%);
-}
-
-/*---------------------------Arrows-----------------------*/
-.arrowLeft {
-  width: 50px;
-  height: 50px;
-  position: absolute;
-  z-index: 2;
-  top: 46.5%;
-  left: 20px;
-  cursor: pointer;
-  background-color: lightgray;
-  margin: 0;
-  padding: 0;
-  -webkit-clip-path: polygon(100% 0%, 75% 50%, 100% 100%, 25% 100%, 0% 50%, 25% 0%);
-          clip-path: polygon(100% 0%, 75% 50%, 100% 100%, 25% 100%, 0% 50%, 25% 0%);
-}
-
-.arrowRight {
-  width: 50px;
-  height: 50px;
-  position: absolute;
-  z-index: 2;
-  top: 46.5%;
-  right: 20px;
-  cursor: pointer;
-  background-color: lightgray;
-  margin: 0;
-  padding: 0;
-  -webkit-clip-path: polygon(75% 0%, 100% 50%, 75% 100%, 0% 100%, 25% 50%, 0% 0%);
-          clip-path: polygon(75% 0%, 100% 50%, 75% 100%, 0% 100%, 25% 50%, 0% 0%);
-}
-
-/*--------------------------------------------------------*/
-/*# sourceMappingURL=gallery.css.map */
