--- src/components/sponsor/Sponsor.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-
-export default class Sponsor extends Component {
-    render() {
-        return (
-            <div>
-                Hi there
-            </div>
-        )
-    }
-}
